+++ src/ExifToolOptions.ts
@@ -1,7 +1,13 @@
 import * as bc from "batch-cluster"
 import { geoTz } from "./GeoTz"
 import { IgnorableError } from "./IgnorableError"
+import { Tags } from "./Tags"
 
+/**
+ * Options for the {@link ExifTool} constructor.
+ *
+ * Defaults are defined in {@link DefaultExifToolOptions}.
+ */
 export interface ExifToolOptions
   extends bc.BatchClusterOptions,
     bc.BatchProcessOptions,
@@ -14,8 +20,8 @@ export interface ExifToolOptions
   maxProcs: number
 
   /**
-   * The maximum number of requests a given ExifTool process will service before
-   * being retired.
+   * The maximum number of requests a given ExifTool process will service
+   * before being retired.
    *
    * Defaults to 500, to balance performance with memory usage.
    */
@@ -23,21 +29,21 @@ export interface ExifToolOptions
 
   /**
    * Spawning new ExifTool processes must not take longer than
-   * `spawnTimeoutMillis` millis before it times out and a new attempt is made.
-   * Be pessimistic here--windows can regularly take several seconds to spin up
-   * a process, thanks to antivirus shenanigans. This can't be set to a value
-   * less than 100ms.
+   * {@link spawnTimeoutMillis} milliseconds before the child process is timed
+   * out and a new attempt is made. Be pessimistic here--windows can regularly
+   * take several seconds to spin up a process, thanks to antivirus
+   * shenanigans. This can't be set to a value less than 100ms.
    *
    * Defaults to 30 seconds, to accommodate slow Windows machines.
    */
   spawnTimeoutMillis: number
 
   /**
-   * If requests to ExifTool take longer than this,
-   * presume the underlying process is dead and we should restart the task. This
-   * can't be set to a value less than 10ms, and really should be set to at more
-   * than a second unless `taskRetries` is sufficiently large or all writes will
-   * be to a fast local disk. Defaults to 10 seconds.
+   * If requests to ExifTool take longer than this, presume the underlying
+   * process is dead and we should restart the task. This can't be set to a
+   * value less than 10ms, and really should be set to at more than a second
+   * unless `taskRetries` is sufficiently large or all writes will be to a
+   * fast local disk. Defaults to 10 seconds.
    */
   taskTimeoutMillis: number
 
@@ -57,8 +63,9 @@ export interface ExifToolOptions
   taskRetries: number
 
   /**
-   * Allows for non-standard paths to ExifTool. Defaults to the perl or windows
-   * binaries provided by `exiftool-vendored.pl` or `exiftool-vendored.exe`.
+   * Allows for non-standard paths to ExifTool. Defaults to the perl or
+   * windows binaries provided by `exiftool-vendored.pl` or
+   * `exiftool-vendored.exe`.
    */
   exiftoolPath: string
 
@@ -76,7 +83,8 @@ export interface ExifToolOptions
    * Should ExifTool use MWG (Metadata Working Group) composite tags for
    * reading and writing tags?
    *
-   * ExifTool recommends this to be set to true.
+   * ExifTool recommends this to be set to true. This defaults to `false` to
+   * maintain consistency with prior versions.
    *
    * Note that this can result in many tag value differences from
    * `ExifTool.read`, and makes `ExifTool.write` write to "synonymous" MWG
@@ -93,8 +101,9 @@ export interface ExifToolOptions
    * If you're using tag values only for human consumption, you may want to
    * leave this blank.
    *
-   * The default includes "*Duration*", "GPSAltitude", "GPSLatitude",
-   * "GPSLongitude", "GPSPosition", and "Orientation".
+   * The default includes "*Duration*", {@link Tags.GPSAltitude},
+   * {@link Tags.GPSLatitude}, {@link Tags.GPSLongitude},
+   * {@link Tags.GPSPosition}, and {@link Tags.Orientation}.
    */
   numericTags: string[]
 
@@ -125,13 +134,14 @@ export interface ExifToolOptions
 
   /**
    * Should we try to backfill timezones for date-times that don't have them?
-   * If set to `true`, and `defaultVideosToUTC` is also `true`, we'll try
-   * backfilling timezones for date-times that are UTC, as well.
+   * If set to `true`, and {@link defaultVideosToUTC} is also `true`, we'll
+   * try backfilling timezones for date-times that are UTC, as well.
    *
    * Setting this to `false` removes **all** timezone inference--only those
-   * date-times with an explicit offset will have a defined timezone. Prior
-   * versions of exiftool-vendored would use the file's `.tz` as a backstop
-   * even if this was set to `false`.
+   * date-times with an explicit offset will have a defined timezone.
+   *
+   * Prior versions of exiftool-vendored would use the file's `.tz` as a
+   * backstop even if this was set to `false`.
    *
    * As of version 23, this now defaults to `true`, as it's more likely to be
    * what people expect.
@@ -139,17 +149,31 @@ export interface ExifToolOptions
   backfillTimezones: boolean
 
   /**
-   * We always look at TimeZone, OffsetTime, TimeZoneOffset, and GPS metadata
-   * to infer the timezone.
+   * We always look at {@link Tags.TimeZone}, {@link Tags.OffsetTime},
+   * {@link Tags.TimeZoneOffset}, {@link Tags.OffsetTimeOriginal},
+   * {@link Tags.OffsetTimeDigitized}, and GPS metadata to infer the timezone.
    *
-   * If these strategies fail, can we infer a timezone from any non-UTC
-   * datestamp whose tag starts with "Creat" and includes an offset? (This
-   * means CreateDateSubSec, CreateDate, and CreationDate will be considered).
+   * If these strategies fail, and this is enabled, we'll try to infer the
+   * timezone from non-UTC datestamps included in the
+   * {@link inferTimezoneFromDatestampTags} value.
    *
    * This defaults to false as it both retains prior behavior and means fewer
    * "fuzzy" heuristics are enabled by default.
    */
   inferTimezoneFromDatestamps: boolean
+
+  /**
+   * This is the list of tag names that will be used to infer the timezone as
+   * a backstop, if no explicit timezone is found in metadata. Note that
+   * datestamps with UTC offsets are ignored, as they are frequently
+   * incorrectly set.
+   *
+   * This setting is only in play if {@link inferTimezoneFromDatestamps} has
+   * been overridden to be `true`.
+   *
+   * This defaults to {@link DefaultCreateDatestampNames}
+   */
+  inferTimezoneFromDatestampTags: (keyof Tags)[]
 
   /**
    * `ExifTool` has a shebang line that assumes a valid `perl` is installed at