+++ src/Struct.ts
@@ -7,7 +7,7 @@ export type Struct = {
 export function isStruct(o: any): o is Struct {
   return (
     o != null &&
-    !Array.isArray(o) &&
+    o.constructor?.name === "Object" &&
     Object.values(o).every((v) => {
       const t = typeof v
       return t === "string" || t === "number" || isDateOrTime(v) || isStruct(v)